backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==25.3.0
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.2
frozenlist==1.6.2
greenlet==3.1.1
groq==0.26.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
marshmallow==3.26.1
motor==3.6.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.5
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
PyJWT==2.10.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.40.0
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.34.2
yarl==1.20.0
zstandard==0.23.0
